Task 0

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	char *buf;
-	int fd;
+	ssize_t fd;
 	ssize_t r, w;
 
 	if (!filename)
